03 VII

+++ src/main/java/com/chrisaraneo/mwl/controller/PlaylistController.java
@@ -10,6 +10,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,8 +22,11 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.chrisaraneo.mwl.exception.BadRequestException;
 import com.chrisaraneo.mwl.exception.ResourceNotFoundException;
+import com.chrisaraneo.mwl.keys.SongPlaylistKey;
 import com.chrisaraneo.mwl.model.Playlist;
+import com.chrisaraneo.mwl.model.RoleName;
 import com.chrisaraneo.mwl.model.SongPlaylist;
 import com.chrisaraneo.mwl.model.User;
 import com.chrisaraneo.mwl.model.extended.PlaylistWithSongs;
@@ -45,12 +52,11 @@ public class PlaylistController {
     
 //    @Autowired
 //    PlaylistRecordRepository songsPlaylistRepository;
-
+    
     @GetMapping("/playlists")
     @Secured({"ROLE_ADMIN", "ROLE_USER"})
     @PreAuthorize("hasRole('USER')")
     public List<Playlist> getAllPlaylists(@CurrentUser UserPrincipal currentUser) {
-    	System.out.println(currentUser.getEmail());
     	Long id = currentUser.getID();
         return playlistRepository.findAllUserPlaylists(id);
     }
@@ -65,9 +71,12 @@ public class PlaylistController {
     	Long userID = currentUser.getID();
     	
     	Playlist playlist = playlistRepository.findUserPlaylist(userID, playlistID);
-    	Set<SongPlaylist> songs = songPlaylistRepository.findAllSongsInPlaylist(playlistID);  
+    	if(playlist == null) {
+    		throw new BadRequestException("This playlist doesn't exist or user is not authorized to GET this playlist.");
+    	}
     	
-        return new PlaylistWithSongs(playlist, songs);
+    	Set<SongPlaylist> songs = songPlaylistRepository.findAllSongsInPlaylist(playlistID);
+		return new PlaylistWithSongs(playlist, songs);
     }
 
     @PostMapping("/playlists")
@@ -89,8 +98,8 @@ public class PlaylistController {
         return null;
     }
     
-    @PutMapping("/playlist/{id}")
-    @Secured("ROLE_USER")
+    @PutMapping("/playlists/{id}")
+    @Secured({"ROLE_ADMIN", "ROLE_USER"})
     @PreAuthorize("hasRole('USER')")
     public Playlist updatePlaylist(
     		@CurrentUser UserPrincipal currentUser,
@@ -115,23 +124,53 @@ public class PlaylistController {
     	return null;
     }
     
-    @DeleteMapping("/playlist/{id}")
-    @Secured("ROLE_ADMIN")
-    @PreAuthorize("hasRole('ADMIN')")
+    private void removeAllSongsFromPlaylist(Playlist playlist) {
+    	Integer playlistID = playlist.getPlaylistID();
+    	// Removing all song-playlist records
+    	Set<SongPlaylist> sp = songPlaylistRepository.findAllSongsInPlaylist(playlistID);
+    	SongPlaylistController spc = new SongPlaylistController();
+    	for(SongPlaylist song : sp) {
+    		Integer songID = song.getSong().getSongID();
+    		Integer track = song.getId().getTrackNumber();
+    		SongPlaylistKey id = new SongPlaylistKey(track, playlist);
+    		songPlaylistRepository.deleteById(id);
+//    		spc.removeSongFromPlaylist(playlistID, songID, track);
+    	}
+    }
+    
+    @DeleteMapping("/playlists/{id}")
+    @Secured({"ROLE_ADMIN", "ROLE_USER"})
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<?> deletePlaylist(
     		@CurrentUser UserPrincipal currentUser,
     		@PathVariable(value = "id") Integer playlistID) {
     	
-    	String email = currentUser.getEmail();
-    	String username = currentUser.getUsername();
-    	Optional<User> user = userRepository.findByUsernameOrEmail(username, email);
-    	if(user.isPresent()) {
-    		Playlist playlist = playlistRepository.findById(playlistID)
-                    .orElseThrow(() -> new ResourceNotFoundException("Playlist", "id", playlistID));
-
-            playlistRepository.delete(playlist);
-
-            return ResponseEntity.ok().build();
+    	Playlist playlist = playlistRepository.findById(playlistID)
+                .orElseThrow(() -> new ResourceNotFoundException("Playlist", "id", playlistID));
+    	
+    	if(currentUser.hasRole((RoleName.ROLE_ADMIN).toString())) {
+    		removeAllSongsFromPlaylist(playlist);
+    		playlistRepository.delete(playlist);
+    		return ResponseEntity.ok().build();
+    	} else {
+    		String email = currentUser.getEmail();
+    		String username = currentUser.getUsername();
+    		Optional<User> user = userRepository.findByUsernameOrEmail(username, email);
+    		if(user.isPresent()) {
+    			User u = playlist.getUser();
+    			if(u != null) {
+    				Optional<User> creator = userRepository.findByUsernameOrEmail(u.getUsername(), u.getEmail());
+    				if(creator.isPresent()) {
+    					if(creator.get().equals(user.get())) {
+    						removeAllSongsFromPlaylist(playlist);
+    						playlistRepository.delete(playlist);
+    						return ResponseEntity.ok().build(); // TODO (?)
+    					} else {
+    						System.out.println("NOT EQUALS");
+    					}
+    				}
+    			}
+    		}
     	}
     	
     	// TODO
